+++ modules/silicon_design/variables.tf
@@ -66,9 +66,13 @@ variable "billing_budget_services" {
 }
 
 variable "billing_budget_notification_email_addresses" {
-  description = "A list of email addresses which will be recieving billing budget notification alerts. A maximum of 5 channels are allowed. {{UIMeta group=0 order=13 updatesafe }}"
-  type        = list(string)
+  description = "A list of email addresses which will be recieving billing budget notification alerts. A maximum of 4 channels are allowed as the first element of `trusted_users` is automatically added as one of the channel. {{UIMeta group=0 order=13 updatesafe }}"
+  type        = set(string)
   default     = []
+  validation {
+    condition     = length(var.billing_budget_notification_email_addresses) <= 4
+    error_message = "Maximum of 4 email addresses are allowed for the budget monitoring channel."
+  }
 }
 
 variable "billing_budget_pubsub_topic" {
@@ -123,6 +127,12 @@ variable "folder_id" {
   description = "Folder ID where the project should be created. It can be skipped if already setting organization_id. Leave blank if the project should be created directly underneath the Organization node. {{UIMeta group=0 order=2 updatesafe }}"
   type        = string
   default     = ""
+}
+
+variable "image_name" {
+  description = "Basename for for the compute and container image. {{UIMeta group=3 order=6 }}"
+  type        = string
+  default     = "silicon-design-ubuntu-2004"
 }
 
 variable "ip_cidr_range" {