+++ CHANGELOG.md
@@ -15,6 +15,21 @@ Here we write upgrading notes for RAD Lab repo. It's a team effort to make them 
  
 ### Fixed
 
+## [9.0.0](https://github.com/GoogleCloudPlatform/rad-lab/releases/tag/v9.0.0) (2022-10-20)
+
+[Full Changelog](https://github.com/GoogleCloudPlatform/rad-lab/compare/v8.0.0...v9.0.0) 
+
+### Added
+- _RAD Lab Module:_ RAD Lab Web Hosting (3-tier architecture) Module [(#79)](https://github.com/GoogleCloudPlatform/rad-lab/pull/79)
+- _RAD Lab Module:_ Added PubSub Topic & Subscription for Billing Budgets. [(#77)](https://github.com/GoogleCloudPlatform/rad-lab/pull/77)
+
+### Changed
+- _RAD Lab Module:_ Using deeplearning images in Silicon Design module. [(#50)](https://github.com/GoogleCloudPlatform/rad-lab/pull/50)
+- _RAD Lab Module:_ Added permission to trusted users/groups for Genomics Cromwell Module. [(#81)](https://github.com/GoogleCloudPlatform/rad-lab/pull/81)
+
+### Fixed
+- Fixed the deployment issue for App Mod Elastic RAD Lab module. [(#74)](https://github.com/GoogleCloudPlatform/rad-lab/pull/74)
+
 ## [8.0.0](https://github.com/GoogleCloudPlatform/rad-lab/releases/tag/v8.0.0) (2022-09-02)
  
 [Full Changelog](https://github.com/GoogleCloudPlatform/rad-lab/compare/v7.0.0...v8.0.0)